fix typo, rename 1317 to 1318
