--- src/leetcode/algorithm_1317/README.md
@@ -1,3 +1,0 @@
-Convert Integer to the Sum of Two No-Zero Integers
-==================================================
-[leetcode](https://leetcode.com/problems/convert-integer-to-the-sum-of-two-no-zero-integers)
