+++ src/leetcode/algorithm_1318/README.md
@@ -1,0 +1,3 @@
+Minimum Flips to Make a OR b Equal to c
+=======================================
+[leetcode](https://leetcode.com/problems/minimum-flips-to-make-a-or-b-equal-to-c)
