+++ src/leetcode/mod.rs
@@ -10,15 +10,19 @@ pub mod algorithm_109;
 pub mod algorithm_110;
 pub mod algorithm_1129;
 pub mod algorithm_1162;
+pub mod algorithm_118;
 pub mod algorithm_12;
 pub mod algorithm_1217;
 pub mod algorithm_1233;
 pub mod algorithm_124;
 pub mod algorithm_125;
 pub mod algorithm_1254;
+pub mod algorithm_1277;
+pub mod algorithm_1304;
 pub mod algorithm_1312;
-pub mod algorithm_1317;
+pub mod algorithm_1318;
 pub mod algorithm_1319;
+pub mod algorithm_1323;
 pub mod algorithm_1345;
 pub mod algorithm_1351;
 pub mod algorithm_1353;
@@ -34,8 +38,10 @@ pub mod algorithm_1456;
 pub mod algorithm_1470;
 pub mod algorithm_1472;
 pub mod algorithm_1491;
+pub mod algorithm_1493;
 pub mod algorithm_1498;
 pub mod algorithm_1502;
+pub mod algorithm_1504;
 pub mod algorithm_152;
 pub mod algorithm_1523;
 pub mod algorithm_1539;
@@ -44,7 +50,9 @@ pub mod algorithm_1572;
 pub mod algorithm_1579;
 pub mod algorithm_1603;
 pub mod algorithm_1675;
+pub mod algorithm_1695;
 pub mod algorithm_1697;
+pub mod algorithm_1717;
 pub mod algorithm_1751;
 pub mod algorithm_1768;
 pub mod algorithm_1799;
@@ -54,23 +62,31 @@ pub mod algorithm_189;
 pub mod algorithm_19;
 pub mod algorithm_190;
 pub mod algorithm_1920;
+pub mod algorithm_1948;
+pub mod algorithm_1957;
 pub mod algorithm_1964;
 pub mod algorithm_198;
 pub mod algorithm_200;
+pub mod algorithm_2044;
 pub mod algorithm_206;
 pub mod algorithm_2099;
+pub mod algorithm_2106;
 pub mod algorithm_211;
 pub mod algorithm_2130;
 pub mod algorithm_2140;
 pub mod algorithm_2163;
 pub mod algorithm_2187;
+pub mod algorithm_2210;
 pub mod algorithm_2218;
+pub mod algorithm_2264;
 pub mod algorithm_228;
 pub mod algorithm_23;
 pub mod algorithm_2300;
 pub mod algorithm_2306;
+pub mod algorithm_231;
 pub mod algorithm_2311;
 pub mod algorithm_2316;
+pub mod algorithm_2322;
 pub mod algorithm_2336;
 pub mod algorithm_2348;
 pub mod algorithm_2360;
@@ -80,25 +96,50 @@ pub mod algorithm_24;
 pub mod algorithm_2402;
 pub mod algorithm_2405;
 pub mod algorithm_2410;
+pub mod algorithm_2411;
+pub mod algorithm_2419;
+pub mod algorithm_2438;
 pub mod algorithm_2439;
 pub mod algorithm_2444;
 pub mod algorithm_2466;
 pub mod algorithm_2477;
 pub mod algorithm_2492;
+pub mod algorithm_2561;
 pub mod algorithm_258;
+pub mod algorithm_2749;
 pub mod algorithm_278;
+pub mod algorithm_2787;
 pub mod algorithm_28;
+pub mod algorithm_3000;
+pub mod algorithm_3021;
+pub mod algorithm_3025;
+pub mod algorithm_3027;
 pub mod algorithm_3136;
+pub mod algorithm_3195;
+pub mod algorithm_3197;
 pub mod algorithm_3201;
 pub mod algorithm_3202;
+pub mod algorithm_326;
 pub mod algorithm_3304;
 pub mod algorithm_3307;
 pub mod algorithm_3330;
 pub mod algorithm_3333;
+pub mod algorithm_3363;
+pub mod algorithm_342;
 pub mod algorithm_3439;
 pub mod algorithm_3440;
+pub mod algorithm_3446;
+pub mod algorithm_3459;
 pub mod algorithm_347;
+pub mod algorithm_3477;
+pub mod algorithm_3479;
+pub mod algorithm_3480;
+pub mod algorithm_3487;
+pub mod algorithm_3495;
 pub mod algorithm_35;
+pub mod algorithm_3516;
+pub mod algorithm_36;
+pub mod algorithm_37;
 pub mod algorithm_4;
 pub mod algorithm_438;
 pub mod algorithm_443;
@@ -108,6 +149,7 @@ pub mod algorithm_46;
 pub mod algorithm_461;
 pub mod algorithm_48;
 pub mod algorithm_49;
+pub mod algorithm_498;
 pub mod algorithm_502;
 pub mod algorithm_53;
 pub mod algorithm_540;
@@ -123,6 +165,7 @@ pub mod algorithm_64;
 pub mod algorithm_652;
 pub mod algorithm_668;
 pub mod algorithm_67;
+pub mod algorithm_679;
 pub mod algorithm_704;
 pub mod algorithm_72;
 pub mod algorithm_721;
@@ -132,11 +175,15 @@ pub mod algorithm_76;
 pub mod algorithm_783;
 pub mod algorithm_79;
 pub mod algorithm_797;
+pub mod algorithm_808;
+pub mod algorithm_837;
 pub mod algorithm_839;
 pub mod algorithm_85;
+pub mod algorithm_869;
 pub mod algorithm_87;
 pub mod algorithm_875;
 pub mod algorithm_881;
+pub mod algorithm_898;
 pub mod algorithm_904;
 pub mod algorithm_91;
 pub mod algorithm_912;
@@ -166,50 +213,3 @@ pub mod unique_binary_search_tree_96;
 pub mod valid_mountain_array_941;
 pub mod valid_parentheses_20;
 pub mod wiggle_subsequence_376;
-pub mod algorithm_1948;
-pub mod algorithm_1957;
-pub mod algorithm_1695;
-pub mod algorithm_1717;
-pub mod algorithm_2322;
-pub mod algorithm_3487;
-pub mod algorithm_3480;
-pub mod algorithm_2210;
-pub mod algorithm_2044;
-pub mod algorithm_2411;
-pub mod algorithm_2419;
-pub mod algorithm_898;
-pub mod algorithm_118;
-pub mod algorithm_2561;
-pub mod algorithm_2106;
-pub mod algorithm_3477;
-pub mod algorithm_3479;
-pub mod algorithm_3363;
-pub mod algorithm_808;
-pub mod algorithm_231;
-pub mod algorithm_869;
-pub mod algorithm_2438;
-pub mod algorithm_2787;
-pub mod algorithm_326;
-pub mod algorithm_2264;
-pub mod algorithm_342;
-pub mod algorithm_1323;
-pub mod algorithm_837;
-pub mod algorithm_679;
-pub mod algorithm_1277;
-pub mod algorithm_1504;
-pub mod algorithm_3195;
-pub mod algorithm_3197;
-pub mod algorithm_1493;
-pub mod algorithm_498;
-pub mod algorithm_3000;
-pub mod algorithm_3459;
-pub mod algorithm_3446;
-pub mod algorithm_3021;
-pub mod algorithm_36;
-pub mod algorithm_37;
-pub mod algorithm_3025;
-pub mod algorithm_3027;
-pub mod algorithm_3516;
-pub mod algorithm_2749;
-pub mod algorithm_3495;
-pub mod algorithm_1304;
